view

+++ src/main/java/com/domain/coin/data/CoinDeleteData.java
@@ -1,11 +1,12 @@
 package com.domain.coin.data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class CoinDeleteData implements Serializable {
 
     private String name;
-    private Double value;
+    private BigDecimal value;
 
     public String getName() {
         return name;
@@ -15,11 +16,11 @@ public class CoinDeleteData implements Serializable {
         this.name = name;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
 }
