+++ src/main/java/com/domain/product/data/ProductGetData.java
@@ -1,17 +1,18 @@
 package com.domain.product.data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class ProductGetData implements Serializable {
     private int id;
     private String name;
     private String code;
-    private Double price;
+    private BigDecimal price;
 
     public ProductGetData() {
     }
 
-    public ProductGetData(int id, String name, String code, Double price) {
+    public ProductGetData(int id, String name, String code, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.code = code;
@@ -42,11 +43,11 @@ public class ProductGetData implements Serializable {
         this.code = code;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
